create stockcontroller

+++ app/Http/Controllers/Api/StockController.php
@@ -16,76 +16,63 @@ class StockController extends Controller
     public function index(Request $request)
     {
         try{
-        $query = Stock::where('agency_id',auth()->user()->agency_id)->with(['agency', 'createdBy', 'user']);
+            $query = Stock::where('agency_id',auth()->user()->agency_id)->with(['agency', 'createdBy', 'user']);
 
-        // Filtres de recherche
-        if ($request->filled('search')) {
-            $search = $request->search;
-            $query->where(function($q) use ($search) {
-                $q->where('name', 'like', "%{$search}%")
-                  ->orWhere('location', 'like', "%{$search}%")
-                  ->orWhere('description', 'like', "%{$search}%");
-            });
+            // Filtres de recherche
+            if ($request->filled('search')) {
+                $search = $request->search;
+                $query->where(function($q) use ($search) {
+                    $q->where('name', 'like', "%{$search}%")
+                    ->orWhere('location', 'like', "%{$search}%")
+                    ->orWhere('description', 'like', "%{$search}%");
+                });
+            }
+
+            if ($request->filled('agency_id')) {
+                $query->where('agency_id', $request->agency_id);
+            }
+
+            if ($request->filled('created_by')) {
+                $query->where('created_by', $request->created_by);
+            }
+
+            if ($request->filled('user_id')) {
+                $query->where('user_id', $request->user_id);
+            }
+
+            // Tri par défaut
+            $query->orderBy('created_at', 'desc');
+
+
+
+            $stocks = $query->paginate(15)->withQueryString();
+
+
+            $agencies = Agency::whereIn('id', Stock::select('agency_id')->distinct()->pluck('agency_id'))->latest()->get();
+            $creators = User::whereIn('id', Stock::select('created_by')->distinct()->pluck('created_by'))->latest()->get();
+            $users    = User::whereIn('id', Stock::select('user_id')->distinct()->pluck('user_id'))->get();
+
+
+            $data = [
+                'stocks' => $stocks,
+                'agencies' => $agencies,
+                'creators' => $creators,
+                'users' => $users,
+            ];
+
+
+            return sendResponse($data,'Stocks  récupérées avec succès');
+
+        }catch(\Throwable $e){
+
+            return sendError('Erreur lors de la récupération des stocks: ' ,500,$e->getMessage());
+
         }
 
-        if ($request->filled('agency_id')) {
-            $query->where('agency_id', $request->agency_id);
-        }
-
-        if ($request->filled('created_by')) {
-            $query->where('created_by', $request->created_by);
-        }
-
-        if ($request->filled('user_id')) {
-            $query->where('user_id', $request->user_id);
-        }
-
-        // Tri par défaut
-        $query->orderBy('created_at', 'desc');
-
-
-
-        $stocks = $query->paginate(15)->withQueryString();
-
-
-        $agencies = Agency::whereIn('id', Stock::select('agency_id')->distinct()->pluck('agency_id'))->latest()->get();
-        $creators = User::whereIn('id', Stock::select('created_by')->distinct()->pluck('created_by'))->latest()->get();
-        $users    = User::whereIn('id', Stock::select('user_id')->distinct()->pluck('user_id'))->get();
-
-
-        $data = [
-            'stocks' => $stocks,
-            'agencies' => $agencies,
-            'creators' => $creators,
-            'users' => $users,
-        ];
-
-
-        return sendResponse($data,'Stocks  récupérées avec succès');
-
-    }catch(\Throwable $e){
-
-        return sendError('Erreur lors de la récupération des stocks: ' ,500,$e->getMessage());
 
     }
 
 
-        // Données pour les filtres
-
-        return view('stock.index', compact('stocks', 'agencies', 'creators', 'users'));
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-        $agencies = Agency::all();
-        $users = User::latest()->get();
-        $stocks = Stock::where('agency_id', Auth::user()->agency_id)->latest()->get();
-
-        return view('stock.create', compact(['agencies', 'users','stocks']));
-    }
 
     /**
      * Store a newly created resource in storage.
@@ -114,33 +101,46 @@ class StockController extends Controller
      */
     public function show(Stock $stock)
     {
-        $stock->load(['agency', 'user', 'createdBy']);
+        try {
+            // $stock->load(['agency', 'user', 'createdBy']);
 
-        $recentProducts = $stock->stockProducts()
-            ->with(['product'])
-            ->orderBy('created_at', 'desc')
-            ->take(5)
-            ->get();
+            $recentProducts = $stock->stockProducts()
+                ->with(['product'])
+                ->orderBy('created_at', 'desc')
+                ->take(1)
+                ->get();
 
+            $recentProducts = $stock->stockProducts()
+                ->with(['product' => function($query) {
+                    $query->select('id', 'name', 'code', 'unit', 'image');
+                }])
+                ->latest()
+                ->limit(5)
+                ->get();
+                //  $recentProducts = $stock->stockProducts()
+                //     ->with(['product'])
+                //     ->take(5)
+                //     ->get();
 
-        $stockProducts = $stock->stockProducts()
-            ->with(['product'])
-            ->get();
+            $proformas = $stock->proformas()
+                ->select('id', 'client', 'total_amount', 'sale_date', 'invoice_type', 'note')
+                ->latest()
+                ->limit(10)
+                ->get();
+            $users = $stock->users;
 
-        return view('stock.show', compact('stock', 'recentProducts', 'stockProducts'));
+            $data = [
+                'stock' => $stock,
+                'recent_products' => $recentProducts,
+                'proformas' => $proformas,
+                'users' => $users
+            ];
+            return sendResponse($data, 'Stock retrieved successfully', 200);
+        } catch (\Throwable $e) {
+            return sendError('Erreur lors de la récupération du stock: ', 500, $e->getMessage());
+        }
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(Stock $stock)
-    {
-        $agencies = Agency::latest()->get();
-        $users = User::latest()->get();
-
-
-        return view('stock.edit', compact('stock', 'agencies', 'users'));
-    }
 
     /**
      * Update the specified resource in storage.
@@ -168,9 +168,33 @@ class StockController extends Controller
      */
     public function destroy(Stock $stock)
     {
-        $stock->delete();
+        try {
+            $stock = Stock::findOrFail($id);
 
-        return redirect()->route('stocks.index')
-                        ->with('success', 'Stock supprimé avec succès.');
+            // Vérifications de sécurité
+            if ($stock->stockProducts()->count() > 0) {
+                return sendError('Impossible de supprimer un stock contenant des produits', 400);
+            }
+
+            $stock->delete();
+
+            return sendResponse(null, 'Stock supprimé avec succès');
+
+        } catch (\Throwable $e) {
+            return sendError('Erreur lors de la suppression', 500, $e->getMessage());
+        }
+    }
+
+    public function detachUser($stockId, $userId)
+    {
+        try {
+            $stock = Stock::findOrFail($stockId);
+            $stock->users()->detach($userId);
+
+            return sendResponse(null, 'Utilisateur désassigné avec succès');
+
+        } catch (\Throwable $e) {
+            return sendError('Erreur lors de la désassignation', 500, $e->getMessage());
+        }
     }
 }
